moved src/utils/middleware + serve + serve-static to new folder
