+++ services/web/serve/dev.js
@@ -1,13 +1,13 @@
 const Koa = require('koa');
 const webpack = require('koa-webpack');
-const webpackConfig = require('./webpack.config');
-const envMiddleware = require('./src/utils/middleware/env');
-const historyMiddleware = require('./src/utils/middleware/history');
-const terminateMiddleware = require('./src/utils/middleware/terminate');
+const webpackConfig = require('../webpack.config');
+const envMiddleware = require('./middlewares/env');
+const historyMiddleware = require('./middlewares/history');
+const terminateMiddleware = require('./middlewares/terminate');
 
 const app = new Koa();
 
-const { BIND_PORT, BIND_HOST, publicEnv } = require('./env');
+const { BIND_PORT, BIND_HOST, publicEnv } = require('../env');
 
 (async () => {
   const webpackMiddleware = await webpack({
@@ -31,6 +31,6 @@ const { BIND_PORT, BIND_HOST, publicEnv } = require('./env');
   app.use(historyMiddleware({ apps: ['/'] }));
   app.use(webpackMiddleware);
   app.listen(BIND_PORT, BIND_HOST, () => {
-    console.info(`Running App on http://${BIND_HOST}:${BIND_PORT}/`);
+    console.info(`Running App on http://${BIND_HOST}:${BIND_PORT}`);
   });
 })();
