+++ services/web/serve/middlewares/assets.js
