+++ services/web/serve/static.js
@@ -2,12 +2,13 @@ const Koa = require('koa');
 const koaLogger = require('koa-logger');
 const koaMount = require('koa-mount');
 const koaBasicAuth = require('koa-basic-auth');
-const envMiddleware = require('./src/utils/middleware/env');
-const assetsMiddleware = require('./src/utils/middleware/assets');
-const historyMiddleware = require('./src/utils/middleware/history');
-const templateMiddleware = require('./src/utils/middleware/template');
-const healthCheckMiddleware = require('./src/utils/middleware/healthCheck');
-const redirectHttpsMiddleware = require('./src/utils/middleware/redirectHttps');
+const envMiddleware = require('./middlewares/env');
+
+const assetsMiddleware = require('./middlewares/assets');
+const historyMiddleware = require('./middlewares/history');
+const templateMiddleware = require('./middlewares/template');
+const healthCheckMiddleware = require('./middlewares/healthCheck');
+const redirectHttpsMiddleware = require('./middlewares/redirectHttps');
 
 const {
   BIND_PORT,
@@ -18,7 +19,7 @@ const {
   HTTP_BASIC_AUTH_USER,
   HTTP_BASIC_AUTH_PASS,
   publicEnv,
-} = require('./env');
+} = require('../env');
 
 const app = new Koa();
 
@@ -48,5 +49,7 @@ app.use(templateMiddleware({ apps: ['/'] }));
 
 app.listen(BIND_PORT, BIND_HOST, (err) => {
   if (err) throw err;
-  console.info(`🐬  Prod App server listening at http://${BIND_HOST}:${BIND_PORT} 🐬\r\n\r\n`);
+  console.info(
+    `🐬  Prod App server listening at http://${BIND_HOST}:${BIND_PORT} 🐬\r\n\r\n`
+  );
 });
